Add is_wrapped call for subprocvecenv and update code for circular imports
