+++ stable_baselines3/common/env_util.py
@@ -6,33 +6,7 @@ import gym
 from stable_baselines3.common.atari_wrappers import AtariWrapper
 from stable_baselines3.common.monitor import Monitor
 from stable_baselines3.common.vec_env import DummyVecEnv, SubprocVecEnv, VecEnv
-
-
-def unwrap_wrapper(env: gym.Env, wrapper_class: Type[gym.Wrapper]) -> Optional[gym.Wrapper]:
-    """
-    Retrieve a ``VecEnvWrapper`` object by recursively searching.
-
-    :param env: Environment to unwrap
-    :param wrapper_class: Wrapper to look for 
-    :return: Environment unwrapped till ``wrapper_class`` if it has been wrapped with it
-    """
-    env_tmp = env
-    while isinstance(env_tmp, gym.Wrapper):
-        if isinstance(env_tmp, wrapper_class):
-            return env_tmp
-        env_tmp = env_tmp.env
-    return None
-
-
-def is_wrapped(env: Union[gym.Env, VecEnv], wrapper_class: Type[gym.Wrapper]) -> bool:
-    """
-    Check if a given environment has been wrapped with a given wrapper.
-
-    :param env: Environment to check
-    :param wrapper_class: Wrapper class to look for
-    :return: True if environment has been wrapped with wrapper_class.
-    """
-    return unwrap_wrapper(env, wrapper_class) is not None
+from stable_baselines3.common.utils import is_wrapped
 
 
 def make_vec_env(
